Added testing to makefile and fixed hashtable and linkedlist
Linkedlist and hashtable had issues regarding a memory leak, but that's
been cleaned up now. Also added the `make test` option to the makefile
that runs all the files in test/ individually. In the future this will be
changed to compile all the files together with a designated main file
that calls all the testing functions.